requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
phonenumbers==9.0.12
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==6.4.0
reportlab==4.4.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
xlsxwriter==3.2.5
